requirements.txt
fastapi
uvicorn[standard]
pytest
jinja2
httpx
PyGithub